+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Todos los cambios notables en este proyecto serán documentados en este archivo.
 El formato está basado en [Keep a Changelog](https://keepachangelog.com/es-ES/1.0.0/)
 y este proyecto se adhiere a [Semantic Versioning](https://semver.org/lang/es/).
 
-## [1.1.1] - 2026-07-27
+## [1.1.3] - 2026-07-28
 ### Añadido
 
 - Nuevo modulo de movimientos
@@ -13,8 +13,11 @@ y este proyecto se adhiere a [Semantic Versioning](https://semver.org/lang/es/).
 - Se pone "Movimimientos en mantenimiento para correccion de bugs"
 - Se añade saldo anterior y saldo actual a la hoja de resumen
 
+
 ### Modificado
 -Se modifican formatos
 -Se añade un totalizado para los abonos y cargos
 - Cambios visuales para app_gui.py
 - Script de back para manejar la lectura de movimientos
+- Se corrige recalculacion erronea al re-organizar Dataframe
+- Se elimina boton de "Reorganizar" y se integra funcion al crear .xlsx/.txt